--- web/src/components/SideMenu/index.css
@@ -1,55 +1,0 @@
-.sidemenu {
-    width: 260px;
-    padding: 10px;
-    border: 1px solid rgba(255, 255, 255, .2);
-    background-color: var(--bg-primary);
-    display: flex;
-    flex-direction: column;
-}
-
-.sidemenu-date-group {
-    color: var(--text-3);
-    text-align: start;
-    margin: 10px;
-    margin-top: 20px;
-}
-
-.sidemenu-button {
-    height: 2rem;
-    margin: 2px;
-    padding: 3px;
-    border: none;
-    border-radius: 5px;
-    transition: ease 0.25s all;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-.sidemenu-button:hover {
-    background-color: var(--bg-secondary);
-}
-
-.sidemenu-button.selected {
-    background-color: var(--bg-secondary);
-}
-
-.sidemenu-bottom {
-    margin-top: auto;
-    width: 90%;
-    border: 1px solid var(--border-color);
-}
-
-.sidemenu-bottom-group {
-    margin: 1rem;
-    display: flex;
-}
-
-.sidemenu-bottom-group-items {
-    margin: 0.5rem;
-}
-
-.link-icon {
-    color: unset;
-    text-decoration: none;
-}
